requirements.txt
streamlit
PyYAML
websocket_client